Move IUnitOfWork to domain project
