+++ source/dotnet/Domain/IUnitOfWork.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Application;
+namespace Energinet.DataHub.Wholesale.Domain;
 
 public interface IUnitOfWork
 {
