+++ source/dotnet/Infrastructure/Persistence/UnitOfWork.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Application;
+using Energinet.DataHub.Wholesale.Domain;
 
 namespace Energinet.DataHub.Wholesale.Infrastructure.Persistence;
